[WIP] dot15d4-frame2: alternative frame representation
A first PoC for an alternative IEEE 802.15.4 frame representation.

In a first step only the "emit" case was implemented. "Emitting" reduces
to transmuting from the typed representation to the same-size u8 array
in the union (zero copy).

The representation is fully typed and occupies always exactly
aMaxPhyPacketSize + driver headroom/tailroom bytes (zero cost).

In a next step, a "parse" feature will be implemented on top of
corresponding variants that allow strong typing on the ingress path.

Signed-off-by: Florian Grandel <fgrandel@code-for-humans.de>