refactor(clock): syntonized -> local clock
We decided that we'll keep syntonization separate of the radio clock
implementation. The radio clock will represent a local clock that needs
to be syntonized externally.

What we called "syntonized" instant/duration so far, therefore actually
is a local radio clock instant/duration.

This change renames the instant/duration accordingly.

Signed-off-by: Florian Grandel <fgrandel@code-for-humans.de>